fix(tel-ui): Fix openmct-lib not found
