+++ telemetry/packages/server/package.json
@@ -37,7 +37,9 @@
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.0",
     "socket.io": "^4.6.2",
+    "socket.io-client": "^4.6.2",
     "winston": "^3.8.2",
+    "winston-transport": "^4.6.0",
     "zod": "^3.21.4"
   },
   "devDependencies": {
