+++ telemetry/packages/ui/app/components/ui/carousel.tsx
@@ -1,17 +1,20 @@
 import * as React from "react"
 import useEmblaCarousel, {
-  type EmblaCarouselType as CarouselApi,
-  type EmblaOptionsType as CarouselOptions,
-  type EmblaPluginType as CarouselPlugin,
+  type UseEmblaCarouselType,
 } from "embla-carousel-react"
 import { ArrowLeft, ArrowRight } from "lucide-react"
 
 import { cn } from "@/lib/utils"
 import { Button } from "@/components/ui/button"
 
+type CarouselApi = UseEmblaCarouselType[1]
+type UseCarouselParameters = Parameters<typeof useEmblaCarousel>
+type CarouselOptions = UseCarouselParameters[0]
+type CarouselPlugin = UseCarouselParameters[1]
+
 type CarouselProps = {
   opts?: CarouselOptions
-  plugins?: CarouselPlugin[]
+  plugins?: CarouselPlugin
   orientation?: "horizontal" | "vertical"
   setApi?: (api: CarouselApi) => void
 }
@@ -200,8 +203,7 @@ const CarouselPrevious = React.forwardRef<
       // @ts-ignore
       ref={ref}
       variant={variant}
-      // @ts-ignore
-      size={size}
+      size={size as "default" | "sm" | "lg" | null | undefined}
       className={cn(
         "absolute  h-8 w-8 rounded-full",
         orientation === "horizontal"
@@ -231,8 +233,7 @@ const CarouselNext = React.forwardRef<
       // @ts-ignore
       ref={ref}
       variant={variant}
-      // @ts-ignore
-      size={size}
+      size={size as "default" | "sm" | "lg" | null | undefined}
       className={cn(
         "absolute h-8 w-8 rounded-full",
         orientation === "horizontal"
