+++ telemetry/packages/ui/app/components/views/debug-view/pod-state-updater.tsx
@@ -1,0 +1,12 @@
+import { PodState } from '@/components/shared/pod-state';
+
+export const PodStateUpdater = ({ podId }: { podId: string }) => {
+  return (
+    <div className="flex gap-2">
+      <PodState podId={podId} />
+      <div>
+        <p>Change</p>
+      </div>
+    </div>
+  );
+};
