+++ telemetry/packages/ui/vite.config.ts
@@ -31,7 +31,7 @@ export default defineConfig(({ command, mode }) => {
       viteStaticCopy({
         targets: [
           {
-            src: '../../node_modules/openmct/dist/*',
+            src: './app/node_modules/openmct/dist/*',
             dest: 'openmct-lib',
           },
         ],
